src/tg/requirements.txt
python-telegram-bot==22.5
json_log_formatter
